requirements-dev.txt
django-cors-headers==3.13.0
coverage==6.4.4